requirements-dev.txt
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pycodestyle==2.9.1
mypy==0.971
black==22.6.0
isort==5.10.1
django-stubs==1.12.0

# for the cli only
click==8.1.3
requests==2.28.1

-r requirements.txt